# swagger.yaml
swagger: "2.0"
info:
  description: "This is a User Role Management API"
  version: "1.0.0"
  title: "Nuvalence User Management API"
basePath: "/api/v2"
...
schemes:
  - "https"
  - "http"
paths:
...
  /user/adjust-roles:
    post:
      tags:
        - "user"
      summary: "Add a role to user"
      operationId: "assignRoleToUser"
      parameters:
        - name: userRole
          in: body
          description: "user object"
          required: true
          schema:
            $ref: "#/definitions/UserRoleDTO"
      responses:
        "201":
          description: "successful operation"
        "400":
          description: "Invalid request body"
        "401":
          description: "You are unauthorized to make this request"
        "404":
          description: "Resource not found"
    delete:
      tags:
        - "user"
      summary: "Delete role from user"
      operationId: "Remove role from user"
      parameters:
        - name: userRole
          in: body
          description: "user object"
          required: true
          schema:
            $ref: "#/definitions/UserRoleDTO"
      responses:
        "200":
          description: "successful operation"
        "401":
          description: "You are unauthorized to make this request"
        "404":
          description: "Resource not found"
...
  /user:
    get:
      tags:
        - "user"
      summary: "Get all users"
      operationId: "getUserList"
      parameters:
        - name: "resource"
          in: "query"
          description: "Name of the resource "
          required: false
          type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/AllUsers"
        "401":
          description: "You are unauthorized to make this request"
        "404":
          description: "User not found"
    post:
      tags:
        - "user"
      summary: "Create a user"
      operationId: "addUser"
      parameters:
        - name: body
          in: body
          description: "user object"
          required: true
          schema:
            $ref: "#/definitions/UserCreationRequest"
      responses:
        "201":
          description: "successful operation"
        "400":
          description: "Invalid request body"
        "401":
          description: "You are unauthorized to make this request"
        "404":
          description: "User not found"
...
definitions:
...
  AllUsers:
    type: array
    items:
      $ref: '#/definitions/UserDTO'
  UserDTO:
    type: object
    required:
      - email
      - id
      - displayName
    properties:
      id:
        type: string
        format: uuid
      displayName:
        type: string
      email:
        type: string
      externalId:
        type: string
      assignedRoles:
        type: array
        items:
          $ref: '#/definitions/RoleDTO'
      preferences:
        type: object
        $ref: '#/definitions/UserPreferenceDTO'
      customFields:
        type: array
        items:
          $ref: '#/definitions/UserCustomFieldDTO'
  RoleDTO:
    type: object
    required:
      - roleName
    properties:
      id:
        type: string
        format: uuid
      roleName:
        type: string
      permissions:
        type: array
        items:
          type: string
  UserRoleDTO:
    type: object
    required:
      - userId
      - roleId
    properties:
      userId:
        type: string
        format: uuid
      roleId:
        type: string
        format: uuid
  UserPreferenceDTO:
    type: object
    required:
      - language
      - communicationPreference
    properties:
      id:
        type: string
        format: uuid
      language:
        $ref: '#/definitions/LanguageDTO'
      communicationPreference:
        type: string
...
  UserCreationRequest:
    type: object
    required:
      - email
      - displayName
    properties:
      displayName:
        type: string
      email:
        type: string
      externalId:
        type: string
      initialRoles:
        type: array
        items:
          $ref: '#/definitions/RoleDTO'
      customFields:
        type: array
        items:
          $ref: '#/definitions/CreateOrUpdateUserCustomFieldDTO'
...
  LanguageDTO:
    type: object
    properties:
      id:
        type: string
        format: uuid
      languageName:
        type: string
        example: "English"
      languageStandardId:
        type: string
        example: "en"
  CustomFieldOptionDTO:
    type: object
    properties:
      id:
        type: string
        format: uuid
      optionValue:
        type: string
      displayText:
        type: string
...
  CreateOrUpdateUserCustomFieldDTO:
    type: object
    required:
      - customFieldId
    properties:
      customFieldId:
        type: string
        format: uuid
      value:
        type: object
  UserCustomFieldDTO:
    type: object
    properties:
      id:
        type: string
        format: uuid
      customFieldId:
        type: string
        format: uuid
      type:
        type: string
        enum: *CUSTOM_FIELD_TYPE
      dataType:
        type: string
        enum: *CUSTOM_FIELD_DATA_TYPE
      name:
        type: string
      displayText:
        type: string
      value:
        type: object
      options:
        type: array
        items:
          $ref: "#/definitions/CustomFieldOptionDTO"